remove unneeded Build.PL
The Build.PL requires Module::Build, which is no longer shipped with
perl core. The convention for CPAN installers is to prefer Build.PL over
Makefile.PL, so this introduces an unneeded prereq.

The Build.PL also lacks some of the metadata that is included in the
Makefile.PL, so if it was used by the author it would produce a release
with less information.

Depending on which file is used for installation, it will produce a
different prereq list. Makefile.PL will include ExtUtils::MakeMaker in
the configure prereqs. This can then result in inconsistent results in a
cpanfile.snapshot, resulting in unneeded churn.

Since the Build.PL serves little use, just remove it to resolve these
inconsistencies and prereq issues.